new: add ``verify`` to enable/disable SSL cert validation.
Note that in python version < 2.7.9, this will fail if ``verify`` is
``True`` as by default, python library would NOT verify SSL certs.
backporting the validation was not included in this version so it'll
just fail with a hopefully clear error message. On version > 2.7.9,
you can disable (or leave enabled) the verification.